split serializer into modules
