+++ src/ui_configuration/generator.rs
@@ -1,8 +1,9 @@
 use crate::dsl::compiler::compile;
 use crate::dsl::compiler::CompilationError;
 use crate::dsl::compiler::CompiledSchema;
-use crate::ui_configuration::json_schema::JsonSchema;
-use crate::ui_configuration::ui_object::UiObject;
+use crate::ui_configuration::JsonSchema;
+use crate::ui_configuration::UiObject;
+
 use serde_json::Map;
 
 pub struct Generator {
