+++ src/ui_configuration/serialization/mod.rs
@@ -1,0 +1,4 @@
+mod json_schema;
+mod object_types;
+mod properties;
+mod ui_object;
