+++ src/ui_configuration/serialization/object_types.rs
@@ -4,58 +4,13 @@ use crate::dsl::object_types::bounds::IntegerObjectBounds;
 use crate::dsl::object_types::bounds::StringLength;
 use crate::dsl::object_types::bounds::StringObjectBounds;
 use crate::dsl::object_types::RawObjectType;
-use crate::dsl::schema::{Property, PropertyList};
+use serde::ser::Error;
+use serde::ser::SerializeMap;
+use serde::Serialize;
+use serde::Serializer;
+use std::string::ToString;
+
 use heck::MixedCase;
-use serde::ser::{Error, Serialize, SerializeMap, Serializer};
-use std::collections::HashMap;
-
-pub fn serialize_property_list<O, E, S>(property_list: &PropertyList, map: &mut S) -> Result<(), E>
-where
-    E: Error,
-    S: SerializeMap<Ok = O, Error = E>,
-{
-    if !property_list.entries.is_empty() {
-        let mut properties_map = HashMap::new();
-        for property in &property_list.entries {
-            properties_map.insert(&property.name, &property.property);
-        }
-        map.serialize_entry("properties", &properties_map)?;
-    };
-
-    let required = &property_list.required_property_names();
-    if !required.is_empty() {
-        map.serialize_entry("required", required)?;
-    }
-
-    let names = &property_list.property_names();
-    if !names.is_empty() {
-        map.serialize_entry("$$order", names)?;
-    }
-    Ok(())
-}
-
-impl Serialize for Property {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut map = serializer.serialize_map(None)?;
-        for title in &self.display_information.title {
-            map.serialize_entry("title", &title)?;
-        }
-
-        for type_spec in &self.type_information {
-            serialize_object_type(&type_spec.inner(), &mut map)?;
-        }
-
-        let property_list = &self.property_list;
-        if property_list.is_some() {
-            serialize_property_list(&property_list.clone().unwrap(), &mut map)?;
-        }
-
-        map.end()
-    }
-}
 
 impl Serialize for EnumerationValue {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
@@ -72,7 +27,7 @@ impl Serialize for EnumerationValue {
     }
 }
 
-fn serialize_object_type<O, E, S>(raw_type: &RawObjectType, map: &mut S) -> Result<(), E>
+pub fn serialize_object_type<O, E, S>(raw_type: &RawObjectType, map: &mut S) -> Result<(), E>
 where
     E: Error,
     S: SerializeMap<Ok = O, Error = E>,
