+++ src/ui_configuration/serialization/ui_object.rs
@@ -2,33 +2,9 @@ use core::borrow::Borrow;
 use crate::dsl::schema::Property;
 use crate::dsl::schema::PropertyList;
 use crate::dsl::schema::SourceSchema;
-use serde_derive::Serialize;
+use crate::ui_configuration::UiObject;
+use crate::ui_configuration::UiObjectProperty;
 use std::collections::HashMap;
-
-#[derive(Serialize)]
-pub struct UiObject<'a>(HashMap<&'a str, UiObjectProperty<'a>>);
-
-impl<'a> UiObject<'a> {
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
-#[derive(Serialize)]
-struct UiObjectProperty<'a> {
-    #[serde(rename = "ui:help")]
-    help: Option<&'a str>,
-    #[serde(rename = "ui:warning")]
-    warning: Option<&'a str>,
-    #[serde(rename = "ui:description")]
-    description: Option<&'a str>,
-}
-
-impl<'a> UiObjectProperty<'a> {
-    pub fn is_empty(&self) -> bool {
-        self.help.is_none() && self.warning.is_none() && self.description.is_none()
-    }
-}
 
 impl<'a> From<&'a SourceSchema> for UiObject<'a> {
     fn from(schema: &'a SourceSchema) -> Self {
